Update and fixes
Expand priority range to 0-6, 6=default
Fix tar, bzip2, text parser
Add support for files larger then 2GB
In text parser switch to 0 priority when at least 32MB is processed (parsing from ~6mb/s to ~40mb/s)

+++ analyzer/parsers/nesparser.hpp
@@ -18,7 +18,7 @@ class NesParser: public Parser {
     uint64_t nesh,nesp,nesc;
     uint64_t info;
     uint32_t buf0;
-    uint32_t i;
+    uint64_t i;
     Filetype type;
     uint64_t jstart, jend, inSize, inpos;
 public:
