+++ analyzer/parsers/uueparser.hpp
@@ -11,7 +11,7 @@ class uueParser: public Parser {
     uint64_t uudstart,uudend,uudline,uudnl,uudlcount,uuc;
     uint64_t info;
     uint32_t buf0, buf1, buf2;
-    uint32_t i;
+    uint64_t i;
     Filetype type;
     uint64_t jstart, jend, inSize, inpos;
 public:
